requirements.txt
Django==6.0
psycopg2-binary
gunicorn
Pillow


